fix test for 386 Windows test

+++ func_test.go
@@ -125,7 +125,7 @@ func TestRegisterFunc_Floats(t *testing.T) {
 func TestRegisterLibFunc_Bool(t *testing.T) {
 	// this callback recreates the state where the return register
 	// contains other information but the least significant byte is false
-	cbFalse := purego.NewCallback(func() uintptr {
+	cbFalse := purego.NewCallback(func() uint64 {
 		return 0x7F5948AE9A00
 	})
 	var runFalse func() bool
